fix(Performance): POC for moving txlist processing from selector to saga
