+++ packages/blockchain-wallet-v4/src/redux/data/btc/sagas.js
@@ -1,9 +1,18 @@
 import { call, put, select, take } from 'redux-saga/effects'
-import { indexBy, length, path, prop } from 'ramda'
+import { indexBy, length, map, path, prop } from 'ramda'
 import * as A from './actions'
 import * as AT from './actionTypes'
 import * as S from './selectors'
 import * as selectors from '../../selectors'
+import Remote from '../../../remote'
+import * as walletSelectors from '../../wallet/selectors'
+import { Wallet, HDAccountList, TXNotes } from '../../../types'
+import { getAddressLabel } from '../../kvStore/btc/selectors'
+import { getBuySellTxHashMatch } from '../../kvStore/buySell/selectors'
+import { getLockboxBtcAccounts } from '../../kvStore/lockbox/selectors'
+import * as transactions from '../../../transactions'
+
+const transformTx = transactions.btc.transformTx
 
 const TX_PER_PAGE = 10
 
@@ -70,7 +79,42 @@ export default ({ api }) => {
       })
       const atBounds = length(data.txs) < TX_PER_PAGE
       yield put(A.transactionsAtBound(atBounds))
-      yield put(A.fetchTransactionsSuccess(data.txs, reset))
+      // PROCESS TXS
+      // Page == Remote ([Tx])
+      // Remote(wallet)
+      const wallet = yield select(walletSelectors.getWallet)
+      const walletR = Remote.of(wallet)
+      // Remote(blockHeight)
+      const blockHeightR = yield select(S.getLatestBlock)
+      // Remote(lockboxXpubs)
+      const accountListR = (yield select(getLockboxBtcAccounts))
+        .map(HDAccountList.fromJS)
+        .getOrElse([])
+      // [Remote([tx])] == [Page] == Pages
+      const state = yield select()
+      const getDescription = (hash, to) =>
+        TXNotes.selectNote(hash, Wallet.selectTxNotes(wallet)) ||
+        getAddressLabel(to, state).getOrElse('')
+
+      const getPartnerLabel = hash => getBuySellTxHashMatch(state, hash)
+
+      // transformTx :: wallet -> blockHeight -> Tx
+      // ProcessPage :: wallet -> blockHeight -> [Tx] -> [Tx]
+      const ProcessTxs = (wallet, block, accountList, txList) =>
+        map(
+          transformTx.bind(
+            undefined,
+            wallet.getOrFail('no_wallet'),
+            block.getOrElse(0),
+            accountList,
+            getDescription,
+            getPartnerLabel
+          ),
+          txList
+        )
+      // ProcessRemotePage :: Page -> Page
+      const page = ProcessTxs(walletR, blockHeightR, accountListR, data.txs)
+      yield put(A.fetchTransactionsSuccess(page, reset))
     } catch (e) {
       yield put(A.fetchTransactionsFailure(e.message))
     }
